+++ src/Cubism/ZBinDumper.h
@@ -41,10 +41,10 @@ void DumpZBin(const TGrid &grid, const int iCounter, const Real t, const string 
 	FILE *file_id;
 	int status;
 
-	static const unsigned int NCHANNELS = Streamer::NCHANNELS;
-	const unsigned int NX = grid.getBlocksPerDimension(0)*B::sizeX;
-	const unsigned int NY = grid.getBlocksPerDimension(1)*B::sizeY;
-	const unsigned int NZ = grid.getBlocksPerDimension(2)*B::sizeZ;
+	static const size_t NCHANNELS = Streamer::NCHANNELS;
+	const size_t NX = grid.getBlocksPerDimension(0)*B::sizeX;
+	const size_t NY = grid.getBlocksPerDimension(1)*B::sizeY;
+	const size_t NZ = grid.getBlocksPerDimension(2)*B::sizeZ;
 
     Real memsize = (NX * NY * NZ * sizeof(Real))/(1024.*1024.*1024.);
     cout << "Allocating " << memsize << " GB of BIN data" << endl;
@@ -52,13 +52,13 @@ void DumpZBin(const TGrid &grid, const int iCounter, const Real t, const string 
 
 	vector<BlockInfo> vInfo_local = grid.getBlocksInfo();
 
-	static const unsigned int sX = 0;
-	static const unsigned int sY = 0;
-	static const unsigned int sZ = 0;
+	static const size_t sX = 0;
+	static const size_t sY = 0;
+	static const size_t sZ = 0;
 
-	static const unsigned int eX = B::sizeX;
-	static const unsigned int eY = B::sizeY;
-	static const unsigned int eZ = B::sizeZ;
+	static const size_t eX = B::sizeX;
+	static const size_t eY = B::sizeY;
+	static const size_t eZ = B::sizeZ;
 
 	sprintf(filename, "%s/%s.zbin", dump_path.c_str(), fullname.c_str());
 
@@ -66,26 +66,26 @@ void DumpZBin(const TGrid &grid, const int iCounter, const Real t, const string 
 
 	header_serial tag;
     fseek(file_id, sizeof(tag), SEEK_SET);
-	for (unsigned int ichannel = 0; ichannel < NCHANNELS; ichannel++)
+	for (size_t ichannel = 0; ichannel < NCHANNELS; ichannel++)
     {
 
 #pragma omp parallel for
-        for(unsigned int i=0; i<vInfo_local.size(); i++)
+        for(size_t i=0; i<vInfo_local.size(); i++)
         {
             BlockInfo& info = vInfo_local[i];
-            const unsigned int idx[3] = {info.index[0], info.index[1], info.index[2]};
+            const size_t idx[3] = {info.index[0], info.index[1], info.index[2]};
             B & b = *(B*)info.ptrBlock;
             Streamer streamer(b);
 
-            for(unsigned int ix=sX; ix<eX; ix++)
+            for(size_t ix=sX; ix<eX; ix++)
             {
-                const unsigned int gx = idx[0]*B::sizeX + ix;
-                for(unsigned int iy=sY; iy<eY; iy++)
+                const size_t gx = idx[0]*B::sizeX + ix;
+                for(size_t iy=sY; iy<eY; iy++)
                 {
-                    const unsigned int gy = idx[1]*B::sizeY + iy;
-                    for(unsigned int iz=sZ; iz<eZ; iz++)
+                    const size_t gy = idx[1]*B::sizeY + iy;
+                    for(size_t iz=sZ; iz<eZ; iz++)
                     {
-                        const unsigned int gz = idx[2]*B::sizeZ + iz;
+                        const size_t gz = idx[2]*B::sizeZ + iz;
 
                         assert((gz + NZ * (gy + NY * gx)) < NX * NY * NZ);
 
@@ -105,7 +105,7 @@ void DumpZBin(const TGrid &grid, const int iCounter, const Real t, const string 
         long local_count = NX * NY * NZ * 1;
         long local_bytes =  local_count * sizeof(Real);
 
-        unsigned int max = local_bytes;
+        size_t max = local_bytes;
         //	int layout[4] = {NCHANNELS, NX, NY, NZ};
         int layout[4] = {NX, NY, NZ, 1};
         long compressed_bytes = ZZcompress((unsigned char *)array_all, local_bytes, layout, &max);	// "in place"
@@ -168,7 +168,7 @@ void ReadZBin(TGrid &grid, const string f_name, const string read_path=".")
     }
 #endif
 
-    for (unsigned int ichannel = 0; ichannel < NCHANNELS; ichannel++)
+    for (size_t ichannel = 0; ichannel < NCHANNELS; ichannel++)
     {
 #if DBG
         printf("compr. size = %ld\n", tag.size[ichannel]); fflush(0);
